# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/federation/logout": {
      "post": {
        "operationId": "FederationController_federationLogout[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Federation-wide logout with back-channel notification of every participating application",
        "tags": [
          "federation"
        ]
      }
    },
    "/federation/logout": {
      "post": {
        "operationId": "FederationController_federationLogout[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Federation-wide logout with back-channel notification of every participating application",
        "tags": [
          "federation"
        ]
      }
    },
    "/v1/federation/clients/{clientId}/refresh": {
      "post": {
        "operationId": "FederationController_refreshClient[0]_v1",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Apply client configuration changes (origins, callbacks, status) immediately",
        "tags": [
          "federation"
        ]
      }
    },
    "/federation/clients/{clientId}/refresh": {
      "post": {
        "operationId": "FederationController_refreshClient[1]",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Apply client configuration changes (origins, callbacks, status) immediately",
        "tags": [
          "federation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Miqaat Identity Federation Service",
    "description": "Central embedded login, federation sessions, RS256 assertions (JWKS), SSO and federation logout. Authentication only - authorization decisions live in the Core Identity Authorization Service.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FederationLogoutDto": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string",
            "example": "rms-web-prod",
            "description": "Required for browser-initiated logout from an application"
          },
          "logout_hint": {
            "type": "string",
            "description": "The `sid` from the assertion the application received"
          },
          "post_logout_redirect_uri": {
            "type": "string",
            "description": "Must exactly match a registered post-logout redirect URI"
          },
          "state": {
            "type": "string"
          },
          "its_id": {
            "type": "string",
            "description": "Service-initiated only: revoke every session of this user"
          },
          "sid": {
            "type": "string",
            "description": "Service-initiated only: revoke this session"
          }
        }
      },
...
    }
  }
}